### 2017-11-12 #447 @leeper
@HenrikBengtsson would you be interested in reviewing this for JOSS?
