### 2017-11-14 #447 @HenrikBengtsson
@leeper, it looks interesting but unfortunately I'm too busy right now.
