### 2017-11-16 #447 @leeper
@sckott would you be interested in reviewing this for JOSS?
