### 2017-11-16 #447 @arfon
> @whedon set @sckott as reviewer

@leeper - there's a slight inconsistency for Whedon here. You need `@whedon assign @sckott as reviewer`
