### 2017-11-16 #463 @whedon
[REVIEW]: simpleCache: R caching for reproducible, distributed, large-scale projects
**Submitting author:** @vpnagraj (<a href="http://orcid.org/0000-0003-0060-566X">VP Nagraj</a>)
**Repository:** <a href="https://github.com/databio/simpleCache" target ="_blank">https://github.com/databio/simpleCache</a>
**Version:** 0.3.1
