### 2017-11-16 #447 @leeper
@sckott Thanks so much! Review is setup over here: https://github.com/openjournals/joss-reviews/issues/463
