### 2017-11-28 databio/simpleCache#27 @sckott
joss review: Two references missing DOIs
https://doi.org/10.1038/nm.4273 and https://doi.org/10.1093/bioinformatics/btv612
