### 2017-11-28 databio/simpleCache#28 @sckott
joss review: contributing guidelines
There is a statement at bottom of README about contributing which is good but it could be made even easier with an issue template, pull request template and a CONTRIBUTING file - e.g. see https://github.com/ropensci/dotgithubfiles
