### 2017-11-28 databio/simpleCache#29 @sckott
joss review: tests could be better
Tests are present in the package, but could be improved.

- test coverage is at 42%, with no coverage in 5 of the files in `R/`. seems smart to test at least minimally all exported functions
- of the tests that do exist, i'm curious if they cover sufficiently complex enough use cases?  I don't know the pkg intimately, so perhaps they do, but at first glance the tests seem rather simple
