### 2017-11-28 databio/simpleCache#30 @sckott
joss review: docs
- I'd suggest including expected types for each parameter, e.g, for `simpleCache` function `cacheName` should be of type `character`, and so on
- I'd make more use of linking to functions, etc. in manual files. e.g,. when you have `simpleCache()` you could instead use `\code{\link{simpleCache}}` or if using markdown then `[simpleCache()]`
- description file: `http://www.github.com/` is now `https://github.com/`
- description file: pretty sure if you use `Authors@R`, then don't use `Maintainer`
- a summary of the exported methods in the pkg would be nice, perhaps in the package level man file (if created) and in the README or vignette.
