### 2017-11-28 #463 @sckott
Nice work @vpnagraj et al. I have a few comments, with linked issues in your GH repo as applicable

- Two of the references in the `paper.md` are missing DOIs (and the papers do have DOIs) - https://github.com/databio/simpleCache/issues/27
- Community guidelines: could be more on this front -  https://github.com/databio/simpleCache/issues/28
- Tests: tests are present, but could be improved - https://github.com/databio/simpleCache/issues/29
- I'm not sure if this was intended, but the package level manual text at the top of the `R/simpleCache.R` file is wrapped into the resulting man file `simpleCache.Rd`. is this intended?  Perhaps if you want a separate pkg level manual file, use `@name simpleCache-package` (I'd recommend that to have package level docs as needed)
- Package docs: some thoughts at https://github.com/databio/simpleCache/issues/30
- Pkg makes extensive use of options in functions. Users can set options outside of the functions in the pkg. I didn't see any docs on setting options via `options` outside of fxns. Do you recommend users do that or not do that?  Curious also why you went with options and not env vars (not leaning one way or other, just curious)?
