### 2017-11-29 #463 @leeper
Thanks for the great review, @sckott! Much appreciated.

@vpnagraj Please try to respond to Scott's points and the specific issues he's raised.
