### 2017-11-30 d702b57
changes to contributing documents as per joss suggestions
