### 2017-12-01 7e740fd
enforceEdgeCharacter is not used anywhere commenting out for now
