### 2017-12-01 6014696
more tests and organization of test code
