+++ tests/testthat/test_all.R
@@ -2,10 +2,24 @@ library(simpleCache)
 
 context("error checking")
 
-test_that("error checking behaves as expected", {
+test_that("notifications and messages as expected", {
+  
+  # message if cache exists
+  simpleCache("normSample", instruction = {rnorm(5e3, 0,1)}, cacheDir = tempdir(), recreate=TRUE)
+  expect_message(simpleCache("normSample", instruction = {rnorm(5e3, 0,1)}, cacheDir = tempdir(), recreate=FALSE, noload = TRUE), "^::Cache exists")
+  deleteCaches("normSample", force = TRUE)
+  
+  # storeCache should not accept non-character cacheName
+  expect_error(storeCache(cacheName = normSample, recreate = TRUE, cacheDir = tempdir()), "storeCache expects the cacheName variable to be a character vector.")
   
   # message when cacheDir isn't defined
   expect_message(simpleCache("normSample", { rnorm(5e3, 0,1) }), regexp = "^No cacheDir specified.")
+  
+  # error when buildDir is empty without instruction
+  expect_error(simpleCache("normSample", cacheDir = tempdir(), buildDir = tempdir(), recreate = TRUE), "::Error::\tNo instruction or RBuild file provided.")
+  
+  # error when buildEnvir includes "instruction"
+  expect_error(simpleCache("normSample", { rnorm(5e3, 0,1) }, buildEnvir = list(instruction="foo"), recreate=TRUE, cacheDir = tempdir()), "Can't provide a variable named 'instruction' in buildEnvir")
   
   # warning when instruction is not expression
   expect_warning(simpleCache("normSample", instruction = "rnorm(5e3, 0,1)", cacheDir = tempdir(), recreate=TRUE))
@@ -16,8 +30,11 @@ test_that("error checking behaves as expected", {
   # error when cacheName is not character
   expect_error(simpleCache(12345, instruction = { rnorm(5e3, 0,1) }, buildDir = NULL, cacheDir = tempdir(), recreate=TRUE))
   
+  # message when return is NULL
+  expect_message(simpleCache("normSample", instruction = {normSample <- NULL}, recreate = TRUE, cacheDir = tempdir()), "NULL value returned, no cache created")
+  
   # we must clean up any temporary caches we make
-  deleteCaches("normSample", force=TRUE)
+  deleteCaches("normSample", force=TRUE, cacheDir = tempdir())
   
 })
 
@@ -62,9 +79,26 @@ test_that("cache can be created without loading", {
   
   expect_null(simpleCache("normSample", { rnorm(5e3, 0,1) }, recreate = TRUE, noload = TRUE, cacheDir = tempdir()))
   
+  expect_true("normSample.RData" %in% listCaches())
+  
   # we must clean up any temporary caches we make
   deleteCaches("normSample", force=TRUE)
   
+})
+
+test_that("object can be stored as cache", {
+
+  normSample2 <<- rnorm(5e3,0,1)
+  
+  expect_message(storeCache("normSample2", cacheDir = NULL, recreate = TRUE), "^You must set global option RCACHE.DIR")
+  
+  expect_message(storeCache("normSample2", cacheDir = tempdir(), recreate = TRUE), "^::Creating cache::")
+  
+  expect_message(storeCache("normSample2", cacheDir = tempdir(), recreate = FALSE), "^::Cache already exists")
+  
+  # we must clean up any temporary caches we make
+  deleteCaches("normSample2", force=TRUE)
+
 })
 
 test_that("option setting works", {
@@ -73,6 +107,7 @@ test_that("option setting works", {
[…]
