### 2017-12-04 bfe1bb4
updates to DESCRIPTION as per JOSS suggestions
