### 2017-12-04 f330063
documentation now formatted with appropriate links out
