+++ man/simpleCache.Rd
@@ -4,7 +4,6 @@
 \name{simpleCache}
 \alias{simpleCache}
 \alias{simpleCache-package}
-\alias{simpleCache}
 \title{Provides intuitive functions for caching R objects, encouraging faster
 reproducible and restartable R analysis}
 \usage{
@@ -33,44 +32,43 @@ for evaluating the code in instruction.}
 load them if they aren't there (like a cache creation loop).}
 
 \item{cacheDir}{The directory where caches are saved (and loaded from).
-Defaults to the global RCACHE.DIR variable}
+Defaults to the global \code{\link[=setCacheDir]{RCACHE.DIR}} RCACHE.DIR variable}
 
-\item{cacheSubDir}{You can specify a subdirectory within the cacheDir
-variable. Defaults to NULL.}
+\item{cacheSubDir}{You can specify a subdirectory within the \code{cacheDir}
+variable. Defaults to \code{NULL}.}
 
 \item{timer}{Report how long it took to create the cache?}
 
 \item{buildDir}{Location of Build files (files with instructions for use If
-the instructions argument is not provided). Defaults to RBUILD.DIR
-global option.}
+the instructions argument is not provided). Defaults to \code{\link[=setCacheBuildDir]{RBUILD.DIR}} global option.}
 
-\item{assignToVariable}{By default, simpleCache assigns the cache to a
-variable named cacheName; you can overrule that here.}
+\item{assignToVariable}{By default, \code{simpleCache} assigns the cache to a
+variable named \code{cacheName}; you can overrule that here.}
 
 \item{loadEnvir}{Into which environment would you like to load the
-variable? Defaults to parent.frame.}
+variable? Defaults to \code{\link[base]{parent.frame}}.}
 
 \item{searchEnvir}{a vector of environments to search for the already loaded
 cache.}
 
-\item{parse}{By default, simpleCache will guess whether you want to parse the
+\item{parse}{By default, \code{simpleCache} will guess whether you want to parse the
 instruction, based on whether it is quoted. You can overwrite the guess
 with this parameter; but this may disappear in the future. In general,
-you should note quote, but use {} around your instructions.}
+you should note quote, but use \{ \} around your instructions.}
 
-\item{nofail}{By default, simpleCache throws an error if the instructions
+\item{nofail}{By default, \code{simpleCache} throws an error if the instructions
 fail. Use this option to convert this error into a warning. No cache will
 be created, but simpleCache will not then hard-stop your processing. This
 is useful, for example, if you are creating a bunch of caches and it's ok
 if some of them do not complete.}
 
-\item{batchRegistry}{A batchtools registry object (built with
-batchtools::makeRegistry()). If provided, this cache will be created on
+\item{batchRegistry}{A \code{batchtools} registry object (built with
+ \code{\link[batchtools]{makeRegistry}}). If provided, this cache will be created on
 the cluster using your batchtools configuration}
 
 \item{batchResources}{A list of variables to provide to batchtools for
-cluster resource managers. Used as the `res` argument to
-batchtools::batchMap()}
+cluster resource managers. Used as the \code{res} argument to
+\code{\link[batchtools]{batchMap}}}
 
 \item{pepSettings}{Experimental untested feature.}
 
@@ -78,8 +76,6 @@ batchtools::batchMap()}
 touch.}
 }
 \description{
-simpleCache provides a function (simpleCache())
-
 Given a unique name for an  R object, and instructions for how to make that
 object, use the simpleCache function to create and cache or load the object.
 This should be used for computations that take a long time and generate a
@@ -89,10 +85,10 @@ troubles if you misuse the caching system. The object should be considered
 static.
 }
 \details{
[…]
