### 2017-12-04 a5643f1
more updates to formatting and linking man pages
