### 2017-12-04 d64c5da
now listing highlights of exported functions in README
