### 2017-12-05 470e33a
deleting enforceEdgeCharacter function altogether
