+++ R/utility.R
@@ -9,25 +9,6 @@
 #
 # These functions should probably remain interior to the package (not exported)
 #
-
-# enforceEdgeCharacter = function(string, prependChar="", appendChar="") {
-# 	if (string=="" | is.null(string)) {
-# 		return(string)
-# 	}
-# 	if(!is.null(appendChar)) {
-# 		if (substr(string,nchar(string), nchar(string)) != appendChar) { # +1 ?
-# 			string = paste0(string, appendChar);
-# 			}
-# 	}
-# 	if (!is.null(prependChar)) {
-# 		if (substr(string,1,1) != prependChar) { # +1 ?
-# 			string = paste0(prependChar, string)
-# 		}
-# 	}
-# 	return(string)
-# }
-
-
 #' Determine if a cache file is sufficiently old to warrant refresh.
 #' 
 #' \code{.tooOld} accepts a maximum cache age and checks for an option with 
