### 2017-12-05 a47c29b
fixed typo and simpleCache description in README
