### 2017-12-06 f53be70
removing obsolete test
