### 2017-12-06 4783cb7
added utils:: to file_test to satisfy devtools::check() note
