### 2017-12-06 5c2c6e5
changed name of viewCacheDirs() to simpleCacheOptions()
