+++ man/viewCacheDirs.Rd
@@ -1,11 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cacheDirectories.R
-\name{viewCacheDirs}
-\alias{viewCacheDirs}
-\title{View cache directories}
-\usage{
-viewCacheDirs()
-}
-\description{
-Views cache directory global variables
-}
