+++ README.md
@@ -87,7 +87,7 @@ change.
 - `listCaches()`: Lists all of the caches available in the `cacheDir`
 - `deleteCaches()`: Deletes cache(s) from the `cacheDir`
 - `setCacheDir()`: Sets a global option for a cache directory so you don't have to specify one in each `simpleCache` call
-- `viewCacheDirs()`: Views all of the `simpleCache` global options that have been set
+- `simpleCacheOptions()`: Views all of the `simpleCache` global options that have been set
 
 ### Contributing
 
