### 2017-12-06 37591a2
updated change log and added loose rd files
