### 2017-12-06 d6d21f0
changed viewCacheDirs name in tests as well
