+++ man/simpleCacheOptions.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/cacheDirectories.R
+\name{simpleCacheOptions}
+\alias{simpleCacheOptions}
+\title{View simpleCache options}
+\usage{
+simpleCacheOptions()
+}
+\description{
+Views simpleCache global variables
+}
