### 2017-12-18 2446a90
Specify arg types; See #30
