### 2017-12-20 69effbc
delay cache recreation
