### 2017-12-20 9e9c38b
remove unnecessary conditional on recreation flag
