### 2017-12-20 06d9224
try slightly longer delay
