### 2017-12-20 da7d11d
explanations and cleaner handling of testdir
