+++ tests/testthat/test_cache_lifespan.R
@@ -42,7 +42,7 @@ my_test_that("Cache file is replaced if no lifespan is specified and recreate=TR
   expect_equal(1, countCacheItems())
   fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
   t0 = file.info(fp)$ctime
-  Sys.sleep(0.25)
+  Sys.sleep(0.25)    # Delay so that our time comparison can work.
   mySimpleCache("testDF", recreate=TRUE, instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
   t1 = file.info(fp)$ctime
@@ -69,7 +69,7 @@ my_test_that("Cache is replaced if older than explicit lifespan", {
   fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
   expect_equal(1, countCacheItems())
   t0 = file.info(fp)$ctime
-  Sys.sleep(0.25)
+  Sys.sleep(0.25)    # Time difference comparison reliability.
   mySimpleCache("testDF", lifespan=0, instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
   t1 = file.info(fp)$ctime
@@ -83,8 +83,8 @@ my_test_that("Cache is replaced if recreate=TRUE even if cache is fresh", {
   fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
   expect_equal(1, countCacheItems())
   t0 = file.info(fp)$ctime
-  Sys.sleep(0.25)
-  mySimpleCache("testDF", recreate=TRUE, lifespan=0, instruction={ buildTestFrame() })
+  Sys.sleep(0.25)    # Time difference comparison reliability.
+  mySimpleCache("testDF", recreate=TRUE, lifespan=1, instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
   t1 = file.info(fp)$ctime
   expect_true(t1 > t0)
