+++ tests/testthat/test_cache_lifespan.R
@@ -84,7 +84,7 @@ my_test_that("Cache is replaced if recreate=TRUE even if cache is fresh", {
   expect_equal(1, countCacheItems())
   t0 = file.info(fp)$ctime
   Sys.sleep(0.25)    # Time difference comparison reliability.
-  mySimpleCache("testDF", recreate=TRUE, lifespan=1, instruction={ buildTestFrame() })
+  mySimpleCache("testDF", recreate=TRUE, lifespan=0, instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
   t1 = file.info(fp)$ctime
   expect_true(t1 > t0)
