### 2017-12-20 bae833f
make direct assertions about time difference
