### 2017-12-20 b32b8db
file existence tests
