+++ tests/testthat/test_cache_lifespan.R
@@ -38,9 +38,11 @@ my_test_that("Cache file isn't replaced if no lifespan is specified and recreate
 # Another sort of control
 my_test_that("Cache file is replaced if no lifespan is specified and recreate=TRUE", {
   expect_equal(0, countCacheItems())
+  fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_false(file_test("-f", fp))
   mySimpleCache("testDF", instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
-  fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_true(file_test("-f", fp))
   t0 = file.info(fp)$ctime
   Sys.sleep(0.25)    # Delay so that our time comparison can work.
   mySimpleCache("testDF", recreate=TRUE, instruction={ buildTestFrame() })
@@ -52,9 +54,11 @@ my_test_that("Cache file is replaced if no lifespan is specified and recreate=TR
 # Specificity
 my_test_that("Cache remains unchanged if younger than explicit lifespan", {
   expect_equal(0, countCacheItems())
+  fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_false(file_test("-f", fp))
   mySimpleCache("testDF", instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
-  fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_true(file_test("-f", fp))
   t0 = file.info(fp)$ctime
   mySimpleCache("testDF", lifespan=0.5, instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
@@ -65,9 +69,11 @@ my_test_that("Cache remains unchanged if younger than explicit lifespan", {
 # Sensitivity
 my_test_that("Cache is replaced if older than explicit lifespan", {
   expect_equal(0, countCacheItems())
-  mySimpleCache("testDF", instruction={ buildTestFrame() })
   fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_false(file_test("-f", fp))
+  mySimpleCache("testDF", instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
+  expect_true(file_test("-f", fp))
   t0 = file.info(fp)$ctime
   Sys.sleep(0.25)    # Time difference comparison reliability.
   mySimpleCache("testDF", lifespan=0, instruction={ buildTestFrame() })
@@ -81,12 +87,10 @@ my_test_that("Cache is replaced if recreate=TRUE even if cache is fresh", {
   expect_equal(0, countCacheItems())
   mySimpleCache("testDF", instruction={ buildTestFrame() })
   fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_true(file_test("-f", fp))
   expect_equal(1, countCacheItems())
   t0 = file.info(fp)$ctime
   Sys.sleep(0.25)    # Time difference comparison reliability.
-  deltaT = difftime(Sys.time(), file.info(fp)$ctime, units="days")
-  expect_true(deltaT > 0)
-  expect_true(deltaT < 1)
   mySimpleCache("testDF", recreate=TRUE, lifespan=0, instruction={ buildTestFrame() })
   expect_equal(1, countCacheItems())
   t1 = file.info(fp)$ctime
