### 2017-12-20 d49c17e
remove unused tmpdir function
