### 2017-12-20 f00ff31
longer sleep for time comparison reliability
