### 2017-12-20 2e824cb
tests regarding option
