+++ tests/testthat/test_cache_lifespan.R
@@ -85,8 +85,9 @@ my_test_that("Cache is replaced if older than explicit lifespan", {
 # Explicit recreate argument trumps cache lifespan to determine recreation.
 my_test_that("Cache is replaced if recreate=TRUE even if cache is fresh", {
   expect_equal(0, countCacheItems())
-  mySimpleCache("testDF", instruction={ buildTestFrame() })
   fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_false(file_test("-f", fp))
+  mySimpleCache("testDF", instruction={ buildTestFrame() })
   expect_true(file_test("-f", fp))
   expect_equal(1, countCacheItems())
   t0 = file.info(fp)$ctime
@@ -95,4 +96,33 @@ my_test_that("Cache is replaced if recreate=TRUE even if cache is fresh", {
   expect_equal(1, countCacheItems())
   t1 = file.info(fp)$ctime
   expect_true(t1 > t0)
+})
+
+my_test_that("simpleCache can pick up option specifying max cache age.", {
+  options(MAX.CACHE.AGE = 0)
+  fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_false(file_test("-f", fp))
+  mySimpleCache("testDF", instruction={ buildTestFrame() })
+  expect_true(file_test("-f", fp))
+  t0 = file.info(fp)$ctime
+  Sys.sleep(1)    # Time difference comparison reliability.
+  mySimpleCache("testDF", instruction={ buildTestFrame() })
+  t1 = file.info(fp)$ctime
+  expect_true(t1 > t0)
+})
+
+my_test_that("Direct lifespan specification is preferred to background option", {
+  options(MAX.CACHE.AGE = 1)
+  fp = file.path(getOption("RCACHE.DIR"), "testDF.RData")
+  expect_false(file_test("-f", fp))
+  mySimpleCache("testDF", instruction={ buildTestFrame() })
+  expect_true(file_test("-f", fp))
+  t0 = file.info(fp)$ctime
+  Sys.sleep(1)
+  mySimpleCache("testDF", instruction={ buildTestFrame() })
+  expect_equal(t0, file.info(fp)$ctime)    # Cache is fresh via MAX.CACHE.AGE.
+  Sys.sleep(1)    # Time difference comparison reliability.
+  mySimpleCache("testDF", lifespan=0, instruction={ buildTestFrame() })
+  t1 = file.info(fp)$ctime
+  expect_true(t1 > t0)    # Cache is stale via lifespan.
 })
