### 2017-12-20 databio/simpleCache#30 @nsheff
Should be all addressed in v0.4.
