### 2017-12-20 databio/simpleCache#27 @nsheff
Should be all addressed in v0.4.
