### 2017-12-20 #463 @nsheff
Hi @leeper and @sckott,

Thanks for the feedback. We've now released [version 0.4](https://github.com/databio/simpleCache/releases), which we think addresses all your comments. Thanks for the review!

To respond to one point here:

>Pkg makes extensive use of options in functions. Users can set options outside of the functions in the pkg. I didn't see any docs on setting options via options outside of fxns. Do you recommend users do that or not do that? Curious also why you went with options and not env vars (not leaning one way or other, just curious)?

We recommend users use the functions to set the options, because these are easier to maintain and change. In a sense, the functions (which are documented) provide a contract with the user; we will honor those functions in the future. The actual option names, and the implementation of how they work, could change at some point, but the functions will remain the same. So, we'd rather just abstract away the actual implementation (whether it uses options, env vars, etc), so that we have room to adjust this as simpleCache advances, without breaking old code. We tried to make this clear with the documentation that explains these setter functions.
