### 2017-12-20 #463 @sckott
explanation of options makes sense
