### 2017-12-21 #463 @leeper
It looks to me like all issues have been addressed. Is that right, @sckott?

@nsheff Can you produce an archive of the repository using, e.g., Zenodo, and let me know the DOI for it?
