### 2018-01-11 #463 @arfon
@vpnagraj @nsheff - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00463 ⚡️ 🚀 💥
